refactor: move definitions from `typval.h` to `typval_defs.h`
Taken from #20187.
